requirements.txt
fastapi
uvicorn
crawlee
playwright
python-dateutil
tabulate 
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
h11==0.14.0
idna==3.7
lxml==5.2.2
playwright==1.45.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
soupsieve==2.5
urllib3==2.2.2
uvicorn==0.30.1
websockets==12.0
greenlet==3.0.3
Werkzeug==3.0.3
zipp==3.19.2 